% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEr-package.R
\docType{package}
\name{DEr-package}
\alias{DEr}
\alias{DEr-package}
\title{DEr: Data Science functions}
\description{
This package provides a set of functions that help with data science workflow.
}
\author{
\strong{Maintainer}: Allan Butler \email{allan.butler@directenergy.com}

}
\keyword{internal}
