% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_functions.R
\name{ImportPeaceData}
\alias{ImportPeaceData}
\title{ImportPeaceData}
\usage{
ImportPeaceData(Platform = "Peace", StartTime = "2018-01-01",
  EndTime = today())
}
\description{
This function allows you import Peace customer data from the ODS.ODS_RO.vw_us_account_base.
}
\examples{
ImportPeaceData()
}
\keyword{Peace}
\keyword{customers}
